[REFACTOR] auth쪽 viewmodel 이전
